Fix instaderails after closing do-less loophole
Changed dotcolor function to say 'black' if asking about a time before
tini. Otherwise for any odom goal starting above zero and do-less goals with
small initial safety buffer, we would think that it was in the red the
day before the goal started, which, by the new derailment criterion
(yesterday = red) would count as derailed on day one!